# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FinaMantic",
        "description": "FinaMantic"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/categories":{
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "All categories list for add goals",
                "description": "All categories list for add goals",
                "operationId": "categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all categories list ",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
        "/users/notification": {
            "get": {
                "tags": ["Profile"],
                "summary": "Notifications",
                "description": "Notifications",
                "operationId": "notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/HeaderVersion"
                    },
                    {
                        "$ref": "#/parameters/HeaderLanguage"
                    },
                    {
                        "$ref": "#/parameters/HeaderPlatform"
                    },
                    {
                        "$ref": "#/parameters/HeaderAuth"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all notifications",
                        "schema": {
                            "$ref": "#/definitions/responsesOk"
                        }
                    },
                    "500": {
                        "description": "Unexpected error.",
                        "schema": {
                            "$ref": "#/definitions/responsesError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responsesOk": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                }
            }
        },
        "responsesError": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "error": {
                    "type": "object"
                }
            }
        },
...
    },
    "parameters": {
        "HeaderVersion": {
            "name": "X-FinaMantic-Version",
            "in": "header",
            "description": "X-FinaMantic-Version",
            "required": true,
            "type": "string",
            "default": "1.0.0"
        },
        "HeaderLanguage": {
            "name": "Accept-Language",
            "in": "header",
            "description": "Accept-Language",
            "required": true,
            "type": "string",
            "default": "en"
        },
        "HeaderPlatform": {
            "name": "X-FinaMantic-Platform",
            "in": "header",
            "description": "android or ios or web",
            "required": true,
            "type": "string",
            "default": "android"
        },
        "HeaderAuth": {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization",
            "required": true,
            "type": "string"
        },
...
    }
}